extracted verify to own routes
